pyproject.toml
[project]
name = "sql"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "pymysql>=1.1.2",
    "python-dotenv>=1.2.1",
]
